implement it (found one potential bug)

+++ core/src/main/java/com/graphhopper/storage/RoutingCHGraph.java
@@ -24,9 +24,9 @@ import com.graphhopper.routing.weighting.Weighting;
 public interface RoutingCHGraph {
     int getNodes();
 
-    int getEdges();
+    long getEdges();
 
-    int getShortcuts();
+    long getShortcuts();
 
     /**
      * Traverses the base edges and shortcuts at a given node. This will only include shortcuts coming from higher
